Add changelog, changeset, and add_table semantics nodes

+++ odb/semantics/relational/elements.hxx
@@ -31,7 +31,7 @@ namespace semantics
     using container::pointer_iterator;
     using compiler::context;
 
-    typedef unsigned int version;
+    typedef unsigned long long version;
 
     //
     //
